// src/app/properties/domain/model/enums/Tag.enum.ts
export enum Tag {
  // MAS AMBIENTES
  COMEDOR_DIARIO = 'COMEDOR_DIARIO',
  DORMITORIO_PRINCIPAL = 'DORMITORIO_PRINCIPAL',
  INGRESO_INDEPENDIENTE = 'INGRESO_INDEPENDIENTE',
  PATIO = 'PATIO',
  SALA_DE_ESTAR = 'SALA_DE_ESTAR',
  AREA_DE_BBQ = 'AREA_DE_BBQ',

  // SERVICIOS
  AREA_DE_LAVANDERIA = 'AREA_DE_LAVANDERIA',
  AREAS_VERDES = 'AREAS_VERDES',
  GUARDIANIA = 'GUARDIANIA',
  PARRILLA = 'PARRILLA',
  SERVICIOS_BASICOS = 'SERVICIOS_BASICOS',
  TELEVISION_POR_CABLE = 'TELEVISION_POR_CABLE',
  GIMNASIO = 'GIMNASIO',
  INTERNET = 'INTERNET',
  LINEA_BLANCA = 'LINEA_BLANCA',
  ILUMINARIAS = 'ILUMINARIAS',
  SALA_Y_COMEDOR = 'SALA_Y_COMEDOR',
  JUEGOS_INFANTILES = 'JUEGOS_INFANTILES',
  SERVICIO_DE_LIMPIEZA = 'SERVICIO_DE_LIMPIEZA',
  SISTEMA_DE_ALARMA = 'SISTEMA_DE_ALARMA',
  USO_COMERCIAL = 'USO_COMERCIAL',
  USO_PROFESIONAL = 'USO_PROFESIONAL',
  VIDEO_VIGILANCIA = 'VIDEO_VIGILANCIA',

  // EXTRAS
  COCINA = 'COCINA',
  CUARTO_DE_SERVICIO = 'CUARTO_DE_SERVICIO',
  ESCUELAS_CERCANAS = 'ESCUELAS_CERCANAS',
  JARDIN = 'JARDIN',
  PET_FRIENDLY = 'PET_FRIENDLY',
  TERRAZA = 'TERRAZA',
  ACABADOS_DE_LUJO = 'ACABADOS_DE_LUJO',
  AMOBLADO = 'AMOBLADO',
  CASETA_DE_GUARDIA = 'CASETA_DE_GUARDIA',
  CENTROS_COMERCIALES_CERCANOS = 'CENTROS_COMERCIALES_CERCANOS',
  CHIMENEA = 'CHIMENEA',
  CLOSET = 'CLOSET',
  COCINA_INTEGRAL = 'COCINA_INTEGRAL',
  EN_CONDOMINIO = 'EN_CONDOMINIO',
  FRENTE_AL_MAR = 'FRENTE_AL_MAR',
  FRENTE_AL_PARQUE = 'FRENTE_AL_PARQUE',
  INTERCOMUNICADOR = 'INTERCOMUNICADOR',
  JACUZZI = 'JACUZZI',
  PARQUES_CERCANOS = 'PARQUES_CERCANOS',
  PISCINA = 'PISCINA',
  SEGURIDAD = 'SEGURIDAD',
  VISTA_A_LA_CIUDAD = 'VISTA_A_LA_CIUDAD',
  VISTA_AL_MAR = 'VISTA_AL_MAR',
  VISTA_AL_PARQUE = 'VISTA_AL_PARQUE',
  WALK_IN_CLOSET = 'WALK_IN_CLOSET',
  ZONA_INDUSTRIAL = 'ZONA_INDUSTRIAL'
}